vectors/vectorizer/track_vectorizer/requirements.txt
langdetect==1.0.9
nltk==3.9.1
numpy==1.26.4
psycopg2-binary==2.9.10
sentence-transformers==4.1.0
